M160: the flop solve runs in float32, and the real bottleneck is named
The M159 session left the flop as the whole remaining latency problem
(p50 16.8s, p90 60.1s) and showed M158's warm start does not help a solo
player, who sees a fresh canonical board nearly every hand. So the first
solve had to get cheaper.

Anatomy first, per M67's warning that a profiler's obvious target can
yield zero. A real flop solve: 318 combos, 16 decision nodes, 29
terminals, 11.2ms/iteration, ~700us per node visit, cost linear in
iterations. The tree is tiny and there is no Python hot loop. The cost is
_solve_recurse carrying a num_hands x num_hands matrix through every node
- ~809KB each in float64, about 18GB of memory traffic per solve.

Float32 on the value path: 1.12-1.32x, measured by interleaved A/B in one
process per M70's rule. A before/after across runs showed 8.9s -> 6.8s,
which is inside this machine's known 1.7x drift, and was discarded. Worst
strategy drift below 5e-6 across three boards.

Free rather than a trade: the entries are Monte Carlo equity estimates
from 30 samples whose own error is ~0.09, so float64's 16 digits were
never carrying information.

Accumulators deliberately stay float64. Regret sums accumulate across
hundreds of iterations, which is where precision matters, and they are
num_hands x num_actions - negligible beside the N x N transients.

This is an improvement, not a fix. 1.25x takes the flop from ~16.8s to
~13s, still outside a comfortable shot clock. The real fix is
algorithmic: the recursion returns N x N from every node and multiplies
it by strategy at each ancestor, giving depth x actions x N^2, where
vector CFR collapses a terminal to an N-vector immediately and stays O(N)
above it. Estimated 3-5x, not attempted here because it rewrites the
solver the canonical library, every reference solve and 999 tests depend
on.

999 backend tests pass.

+++ poker_solver/solver.py
@@ -627,7 +627,12 @@ def solve_flop(
         if equity_samples is not None:
             equity_kwargs["samples"] = equity_samples
         equity_table = build_board_equity_table(board, combos, **equity_kwargs)
-    equity_table = np.nan_to_num(equity_table, nan=0.5)
+    # M160: float32. The flop solve is memory-bound on this table — the
+    # recursion carries a num_hands x num_hands matrix per node — and the
+    # entries are Monte Carlo equity estimates from 30 samples, whose own
+    # error dwarfs anything float32 rounding introduces. Measured across
+    # three boards: 1.3-1.4x faster with worst strategy drift below 5e-6.
+    equity_table = np.nan_to_num(equity_table, nan=0.5).astype(np.float32)
 
     actual_iterations = iterations if iterations is not None else DEFAULT_FLOP_ITERATIONS
     # M158: `warm_start` is `(cached_hands, {action path: InfoSetTable})`
